Use EveryPolitician rather than GitHub webhooks
This changes the countries.json rebuilder to use webhooks from the
EveryPolitician webhook manager. Previously the webhook-manager didn't
have webhooks for open or sync events, which is what this needs, now it
supports those we can use that and have more control over when the app
is triggered, rather than having GitHub ping us any time a pull request
changes in any way.